+++ HU2/MCU/MCU/UI.h
@@ -41,7 +41,7 @@ volatile enum _error _errorcode;
 extern volatile uint16_t gas1perc, gas2perc, brakeperc, predistimer, flowleft, flowright, templeft, tempright, rpm_fl, rpm_fr, rpm_bl, rpm_br;
 extern volatile uint8_t shutdownon, ams_shutdown, imd_shutdown;
 extern volatile uint8_t selsetting, ischanging, vsettings[SETTINGS_COUNT];
-extern volatile int16_t steerpos;
+extern volatile double steerpos;
 extern volatile uint8_t dt_engv, pump_pwm;
 extern uint16_t boot_count;
